requirements.txt
google-cloud-storage
google-cloud-bigquery
requests
python-dotenv
pandas
pytest
responses